p4i integration Integrate change 61674 from 3.0 to main. Description for change 61674:
: Change 61674 by vohi@vohi-hokkaido-nt on 2002/04/15 09:35:35
: 
: 	No warnings.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/examples/demo/frame.cpp#8 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 61675]

+++ examples/demo/frame.cpp
@@ -233,7 +233,7 @@ void Frame::updateTranslators()
 void
 Frame::setCurrentCategory( QWidget *w )
 {
-    for(int i = 0; i < categories->count(); i++) {
+    for(int i = 0; i < (int)categories->count(); i++) {
 	QListBoxItem *item = categories->item(i);
 	if(((CategoryItem*)item)->widget() == w) {
 	    categories->setCurrentItem( i );
